requirements.txt
djangorestframework
djangorestframework-simplejwt
django-model-utils